web_scraper_api/requirements.txt
pydantic~=2.10.6
python-dotenv~=1.0.1
fastapi~=0.115.8
pytz~=2025.1
selenium~=4.29.0
uvicorn~=0.34.0